+++ serde_json_path_core/CHANGELOG.md
@@ -8,9 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 # Unreleased
 
 - **breaking**: ensure integers used as indices are within the [valid range for I-JSON][i-json-range] ([#98])
+- **internal**: remove use of `once_cell` and use specific versions for crate dependencies ([#105])
 
 [#98]: https://github.com/hiltontj/serde_json_path/pull/98
 [i-json-range]: https://www.rfc-editor.org/rfc/rfc9535.html#section-2.1-4.1
+[#105]: https://github.com/hiltontj/serde_json_path/pull/105
 
 # 0.1.6 (3 March 2024)